packages/data_ml/pyproject.toml
[project]
name = "data-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "convex>=0.7.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "mypy>=1.19.1",
    "torch>=2.11.0",
]

[tool.mypy]
python_version = "3.14"
pretty = true
show_error_codes = true
show_column_numbers = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_any_generics = true
strict_equality = true
no_implicit_reexport = true
ignore_missing_imports = true
follow_imports = "normal"
exclude = [
    "^\\.venv/",
]

[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "pandas.*",
    "sklearn.*",
    "convex.*",
    "pytest.*",
    "torch.*",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-cov>=7.1.0",
]
